% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edicio.R
\name{actualitzaInformesCarregats}
\alias{actualitzaInformesCarregats}
\title{DEPRECATED: Actualitza els informes de les edicions carregades}
\usage{
actualitzaInformesCarregats(
  arrelProjecte,
  esborraInformesDesactualitzats = FALSE
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta \code{edicions} i buscarà els informes i revisions a partir d'aquest camí.}

\item{esborraInformesDesactualitzats}{si és \code{TRUE}, elimina els informes desactualitzats per tornar-los a generar de nou.
Altrament, elimina els objectes dels informes que han estat actualitzats.}
}
\value{
Camins dels informes actualitzats.
}
\description{
Useu la funció després de carregar a OSM els fitxers d'edicions. Arxiva els fitxers d'edicions a «$arrelProjecte/edicions/FET» i
 actualitza els informes eliminant els casos ja editats o eliminant els fitxers desactualitzats.
Mou els informes actualitzats a edicions/FET i arxiva els informes originals a ANTIC/.
}
