% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikimedia.R
\name{afegeix_dades_wikidata}
\alias{afegeix_dades_wikidata}
\title{Afegeix dades wikidata}
\usage{
afegeix_dades_wikidata(x)
}
\arguments{
\item{x}{Un `data.frame` amb una columna `wikidata`.}
}
\value{
Retorna `x` amb les columnes `noms_wd` (títol de viquipèdia, etiquetes i àlies), `ca.viquipedia` (títol de
  viquipèdia), `wikidata_tipus` (etiquetes de tots els tipus d'instàncies de wikidata (P31)).
}
\description{
Afegeix dades wikidata
}
