% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte_LangToolsOSM.R
\name{bdRevisions}
\alias{bdRevisions}
\title{Base de dades de revisions}
\usage{
bdRevisions(arrelProjectes)
}
\arguments{
\item{arrelProjectes}{camí a partir del qual es cercaran recursivament les revisions fetes.}
}
\value{
Retorna un \code{data.frame} amb totes les revisions fetes que contenen valors per \code{name:ca} o \code{alt_name:ca}.
}
\description{
Genera revisions per casos que només es diferencien per si els noms comencen amb majúscules o minúscules, mantenint el format original.
}
