% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informesPPCC.R
\name{cami_fitxerPPCC}
\alias{cami_fitxerPPCC}
\title{Camins de fitxers d'un projecte de PPCC}
\usage{
cami_fitxerPPCC(
  divisions,
  arrelProjecte,
  tipus = c("informe", "revisio", "edicio"),
  sufixFitxers,
  format = c("RData", "tsv")
)
}
\arguments{
\item{divisions}{Un data frame com `municipis`, `comarques`, `territoris` o `PPCC`}

\item{arrelProjecte}{Camí a l'arrel del projecte.}

\item{tipus}{description}

\item{sufixFitxers}{Text afegir com a sufix al nom dels fitxers dels informes (`arrelProjecte/informe-Regio-comarca$sufixFitxer$.tsv`).}

\item{format}{Com es desaran els resultats del projecte. Per defecte, "RData".}
}
\description{
Camins de fitxers d'un projecte de PPCC
}
\examples{
cami_fitxerPPCC(divisions = comarques, arrelProjecte = "exemple", sufixFitxers = "_name:ca")
}
