% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informe.R
\name{descarta_objectes_sense_traduccions}
\alias{descarta_objectes_sense_traduccions}
\title{Descarta objectes sense traduccions dels informes}
\usage{
descarta_objectes_sense_traduccions(informes)
}
\arguments{
\item{informes}{Camins dels informes o una llista amb data.frames d'informes.}
}
\value{
Retorna invisiblement els camins dels informes modificats.
}
\description{
Sobreescriu els informes descarant-ne els objectes (files) sense traduccions de wikidata.
}
