% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edicio.R
\name{envia_edicions}
\alias{envia_edicions}
\title{Envia les edicions i actualitza els informes de les edicions carregades}
\usage{
envia_edicions(
  edicions,
  arrelProjecte,
  comentaris,
  hashtags = "#toponimsCat",
  ...,
  esborraInformesDesactualitzats = FALSE,
  xerraire = FALSE
)
}
\arguments{
\item{edicions}{Una llista de taules de canvis o els camins de fitxers que els continguin (`.osc` o `.RData`).}

\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta `edicions` i
buscarà els informes i revisions a partir d'aquest camí.}

\item{comentaris}{Vector de comentaris corresponent a cada conjunt de canvis.}

\item{hashtags}{Etiquetes dels conjunts de canvis.}

\item{...}{Paràmetres per [osmapiR::osm_create_changeset()]. Permet afegir etiquetes als conjunts de canvis
(p.ex. `source`).}

\item{esborraInformesDesactualitzats}{si és `TRUE`, elimina els informes desactualitzats per tornar-los a generar de
nou. Altrament, elimina els objectes dels informes que han estat actualitzats.}

\item{xerraire}{Si és `TRUE`, mostra missatges amb el conjunt de canvis carregat i comentari associat.}
}
\value{
Camins dels informes actualitzats.
}
\description{
Arxiva els fitxers d'edicions a `$arrelProjecte/edicions/FET` i actualitza els informes eliminant els casos ja
carregats o eliminant els informes desactualitzats (arxivatz a ``$arrelProjecte/informes/ANTIC``).
}
