% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_cat_render-filtres.r
\name{filtres_osm_cat}
\alias{filtres_osm_cat}
\title{Filtres dels objectes renderitzats a openstreetmap.cat}
\usage{
filtres_osm_cat()
}
\description{
Filtres extrets de <https://github.com/osm-bzh/osmbr-mapstyle/blob/master/osm_ca.yml>
}
