% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revisio.R
\name{genera_revisions}
\alias{genera_revisions}
\alias{genera_revisio}
\alias{genera_revisions_regex_name}
\alias{genera_revisions_regex_nomsWD}
\title{Genera fitxers de revisió d'informes}
\usage{
genera_revisions(
  informes,
  arrelProjecte,
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "noms_wd",
    "ca.viquipedia", "wikidata_tipus", "wikidata")
)

genera_revisio(
  informe,
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "noms_wd",
    "ca.viquipedia", "wikidata_tipus", "wikidata")
)

genera_revisions_regex_name(
  informes,
  arrelProjecte,
  cerca,
  substitueix,
  revisioUnificada = FALSE,
  nomFitxerUnificat = "revisio-UNIFICADA.tsv",
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "noms_wd",
    "ca.viquipedia", "wikidata_tipus", "wikidata")
)

genera_revisions_regex_nomsWD(
  informes,
  arrelProjecte,
  cerca = " \\\\(.+\\\\)",
  substitueix = "",
  ometSenseTraduccions = TRUE,
  revisioUnificada = FALSE,
  nomFitxerUnificat = "revisio-UNIFICADA.tsv",
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "noms_wd",
    "ca.viquipedia", "wikidata_tipus", "wikidata")
)
}
\arguments{
\item{informes}{camí dels informes pels que es vol generar fitxers de revisió.}

\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta `revisions`.}

\item{filtres}{una llista amb els elements amb noms de columnes dels informes que serveixen per filtrar casos. Cada element és una funció que retorna un vector amb `FALSE` o `TRUE` per cada element de la columna que es passa com a paràmetre.}

\item{campsUnics}{camps que defineixen combinacions úniques dels informes que formaran els fitxers de revisió. Per defecte `name`, `name:ca`, `alt_name:ca`, `alt_name`, `noms_wd`, `ca.viquipedia`, `wikidata_tipus`, `wikidata` si existeixen.}

\item{cerca}{expressió regular del patró de cerca (\link{regex}).}

\item{substitueix}{text a substituir.}

\item{revisioUnificada}{si és `TRUE`, genera un unic fitxer de revisió. Sinó, genera un fixer de revisió per cada informe.}

\item{nomFitxerUnificat}{nom del fitxer de revisió unificat quan `revisioUnificada=TRUE`. Si `revisioUnificada=FALSE`, els fitxers de revisió seran `gsub("^informe", "revisio", informes)`.}

\item{ometSenseTraduccions}{si és `TRUE`, descarta els objectes sense nom en català a wikidata.}
}
\value{
Retorna els camins dels fitxers de revisió generats.
}
\description{
Els fitxers de revisió d'informes contenen els casos a revisar amb combinacions úniques de
les etiquetes `name`, `name:ca`, `alt_name:ca`, `alt_name`,
`noms_wd`, `ca.viquipedia` i `wikidata`.
Cal corregir els casos de `name:ca` i `alt_name:ca` incorrectes i esborrar-los o deixar-los en blanc si no és clar. Un cop revisat,
moveu els fitxers de revisions a «arrelProjecte/revisio/FET/» i prepareu les edicions a OSM amb la funció [preparaEdicions()].
}
\examples{
genera_revisions_regex_name(
  informes = dir("PPCC/calle-carrer/informes", "\\\\.tsv$", full.names = TRUE),
  arrelProjecte = "PPCC/calle-carrer",
  cerca = "([Cc])alle ",
  substitueix = "\\\\1arrer ", revisioUnificada = FALSE
)

genera_revisions_regex_nomsWD(
  informes = dir("exotopònims/Grècia/informes", "\\\\.tsv$",
    full.names = TRUE
  ),
  arrelProjecte = "exotopònims/Grècia",
  cerca = " \\\\(.+\\\\)",
  substitueix = "", revisioUnificada = FALSE
)
}
\seealso{
\link{gsub}
}
