% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte_LangToolsOSM.R
\name{generaInforme}
\alias{generaInforme}
\title{Genera informe}
\usage{
generaInforme(
  arrelProjecte,
  fitxerInforme,
  filtreArea,
  filtreObjectes,
  consulta,
  actualitzaFitxer = FALSE
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta \code{informes}.}

\item{fitxerInforme}{vector amb els camins de fitxers d'informe.}

\item{filtreArea}{filtre d'àrea per la consulta d'Overpass.}

\item{filtreObjectes}{filtre d'etiquetes d'objectes d'OSM per la consulta d'Overpass.}

\item{consulta}{consulta completa d'Overpass. Si existeix el paràmetre, ignora \code{filtreArea} i \code{filtreObjectes}.}

\item{actualitzaFitxer}{si és \code{TRUE} i ja existeix el fitxer d'informe, el mou a la carpeta «ANTIC».}
}
\value{
Ordre per generar l'informe amb \code{write_osm_objects_report} de \href{https://github.com/OSM-Catalan/LangToolsOSM}{LangToolsOSM}.
}
\description{
Genera un informe amb tots els objectes d'OSM segons l'àrea i els filtres especificats.
}
\examples{
ordre0<- generaInforme(arrelProjecte="exotopònims/Rússia/",
                       fitxerInforme="informe-Rússia.tsv",
                       filtreArea="['name:ca'='Rússia'][admin_level=3]",
                       filtreObjectes="nwr[wikidata][!'name:ca']")
ordre1<- generaInforme(arrelProjecte="ppcc/calle-carrer/",
                       fitxerInforme="informe-Calle_carrer-Alacant.tsv",
                       filtreArea="['name:ca'='Alacant'][admin_level=7]",
                       filtreObjectes="nwr[name~'^[Cc]alle '][!'name:ca']")
\dontrun{
# Crida les ordres de LantToolsOSM (cal que estigui instal·lat a l'entorn Python configurat a R)
  system(ordre0)
  system(ordre1)
}
}
