% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informesPPCC.R
\name{generaInformesPPCC}
\alias{generaInformesPPCC}
\title{Genera informes pels  PPCC}
\usage{
generaInformesPPCC(
  arrelProjecte,
  filtre,
  actualitzaInformes = FALSE,
  sufixFitxers = "",
  divisions = toponimsCat::comarques,
  coordenades = FALSE
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació dels informes serà la subcarpeta \code{informes}.}

\item{filtre}{filtre d'etiquetes d'objectes d'OSM per la consulta d'Overpass.}

\item{actualitzaInformes}{si és \code{TRUE} i ja existeix el fitxer d'informe, el mou a la carpeta «ANTIC».}

\item{sufixFitxers}{text afegir com a sufix al nom dels fitxers dels informes («arrelProjecte/informe-Regio-comarca$sufixFitxer$.tsv»).}

\item{divisions}{\code{data.frame} amb informació de les \code{\link{territoris}}, \code{\link{comarques}} o \code{\link{municipis}}. Per defecte, comarques.}

\item{coordenades}{si és \code{TRUE}, els informes inclouen les columnes \code{latitude} i \code{longitude}. Per defecte, \code{FALSE}.}
}
\value{
Retorna la taula de les divisions amb els nous camps \code{cmd}, que conté l'ordre per generar els informes amb
 \code{write_osm_objects_report} de \href{https://github.com/OSM-Catalan/LangToolsOSM}{LangToolsOSM},
 i el camp \code{informe}, que conté el camí del fitxer de l'informe que es crearà.
}
\description{
Generar informes per territoris, comarques o municipis dels Països Catalans amb tots els objectes d'OSM segons els filtres especificats. Només inclou els municipis i zones catalanoparlants de les comarques bilingües.
}
\examples{
ordres<- generaInformesPPCC(arrelProjecte="PPCC/calle-carrer",
                            filtre="nwr[name~'^[Cc]alle'][!'name:ca']",
                            sufixFitxers="_name-calle")
\dontrun{
# Crida les ordres de LantToolsOSM (cal que estigui instal·lat a l'entorn Python configurat a R)
  for (i in 1:length(cmd)){
    message(i, " / ", length(cmd), "\t", cmd[i])
    system(ordres)
  }
}
}
