% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte_LangToolsOSM.R
\name{generaRevisions}
\alias{generaRevisions}
\alias{generaRevisions_regexName}
\alias{generaRevisions_regexTranslations}
\title{Genera fitxers de revisió d'informes}
\usage{
generaRevisions_regexName(
  informes,
  arrelProjecte,
  cerca,
  substitueix,
  revisioUnificada = FALSE,
  nomFitxerUnificat = "revisio-UNIFICADA.tsv",
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "translations",
    "ca.wikipedia_page", "wikidata_type", "wikidata_id")
)

generaRevisions_regexTranslations(
  informes,
  arrelProjecte,
  cerca = " \\\\(.+\\\\)",
  substitueix = "",
  ometSenseTraduccions = TRUE,
  revisioUnificada = FALSE,
  nomFitxerUnificat = "revisio-UNIFICADA.tsv",
  filtres,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "translations",
    "ca.wikipedia_page", "wikidata_type", "wikidata_id")
)
}
\arguments{
\item{informes}{camí dels informes pels que es vol generar fitxers de revisió.}

\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta \code{revisions}.}

\item{cerca}{expressió regular del patró de cerca (\link{regex}).}

\item{substitueix}{text a substituir.}

\item{revisioUnificada}{si és \code{TRUE}, genera un unic fitxer de revisió. Sinó, genera un fixer de revisió per cada informe.}

\item{nomFitxerUnificat}{nom del fitxer de revisió unificat quan \code{revisioUnificada=TRUE}. Si \code{revisioUnificada=FALSE}, els fitxers de revisió seran \code{gsub("^informe", "revisio", informes)}.}

\item{filtres}{una llista amb els elements amb noms de columnes dels informes que serveixen per filtrar casos. Cada element és una funció que retorna un vector amb `FALSE` o `TRUE` per cada element de la columna que es passa com a paràmetre.}

\item{campsUnics}{camps que defineixen combinacions úniques dels informes que formaran els fitxers de revisió. Per defecte \code{name}, \code{name:ca}, \code{alt_name:ca}, \code{alt_name}, \code{translations}, \code{ca.wikipedia_page}, \code{wikidata_type}, \code{wikidata_id} si existeixen.}

\item{ometSenseTraduccions}{si és \code{TRUE}, descarta els objectes sense nom en català a wikidata.}
}
\value{
Retorna els camins dels fitxers de revisió generats.
}
\description{
Els fitxers de revisió d'informes contenen els casos a revisar amb combinacions úniques de
 les etiquetes \code{«name»}, \code{«name:ca»}, \code{«alt_name:ca»}, \code{«alt_name»},
 \code{«translations»}, \code{«ca.wikipedia_page»} i \code{«wikidata_id»}.
 Cal corregir els casos de \code{«name:ca»} i \code{«alt_name:ca»} incorrectes i esborrar-los o deixar-los en blanc si no és clar. Un cop revisat,
 moveu els fitxers de revisions a «arrelProjecte/revisio/FET/» i prepareu les edicions a OSM amb la funció \code{\link{preparaEdicions}}.
}
\examples{
generaRevisions_regexName(informes=dir("PPCC/calle-carrer/informes", "\\\\.tsv$", full.names=TRUE),
                          arrelProjecte="PPCC/calle-carrer",
                          cerca="([Cc])alle ",
                          substitueix="\\\\1arrer ", revisioUnificada=FALSE)

generaRevisions_regexTranslations(informes=dir("exotopònims/Grècia/informes", "\\\\.tsv$",
                                               full.names=TRUE),
                                  arrelProjecte="exotopònims/Grècia",
                                  cerca=" \\\\(.+\\\\)",
                                  substitueix="", revisioUnificada=FALSE)
}
\seealso{
\link{gsub}
}
