% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informe.R
\name{obte_informe}
\alias{obte_informe}
\title{Obté informe}
\usage{
obte_informe(
  arrelProjecte,
  fitxerInforme,
  consulta,
  actualitzaInforme = FALSE,
  consulta_wikidata = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta `informes`.}

\item{fitxerInforme}{Nom de fitxer de l'informe (amb extensió `.RData` o `.tsv`). Camí absolut si no hi ha el
paràmetre `arrelProjecte`.}

\item{consulta}{Consulta d'Overpass com a cadena.}

\item{actualitzaInforme}{si és `TRUE` i ja existeix el fitxer d'informe, el mou a la carpeta «ANTIC».}

\item{consulta_wikidata}{Si és `TRUE`, afegeix columnes amb informació extreta de wikidata (vegeu
[afegeix_dades_wikidata()]).}

\item{verbose}{Si és `TRUE`, treu missatges si el fitxer ja existeix.}
}
\value{
Retorna un `data.frame` amb les dades de l'informe i escriu el resultat a `fitxerInforme` si no existeix el
  fitxer o `actualitzaInforme`.
}
\description{
Genera un informe amb tots els objectes d'OSM segons l'àrea i els filtres especificats.
}
\examples{
consulta <- osmdata::opq(
  bbox = "relation(id:190090)",
  osm_types = "nwr",
  out = "tags",
  timeout = 1000
) |>
  osmdata::add_osm_features(features = filtres_osm_cat(), value_exact = FALSE) |>
  osmdata::add_osm_feature(key = c("name", "!name:ca", "wikidata")) |>
  osmdata::opq_string()
informe <- obte_informe(
  arrelProjecte = "exotopònims/Rússia/",
  fitxerInforme = "informe-Rússia",
  consulta = consulta
)
}
