% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informesPPCC.R
\name{obte_informesPPCC}
\alias{obte_informesPPCC}
\title{Genera informes pels  PPCC}
\usage{
obte_informesPPCC(
  arrelProjecte,
  etiquetes = filtres_osm_cat(),
  claus = c("name", "!name:ca"),
  actualitzaInformes = FALSE,
  sufixFitxers = "",
  divisions = toponimsCat::comarques,
  consulta_wikidata = FALSE,
  coordenades = FALSE,
  data = NULL,
  data2 = NULL,
  adiff = FALSE,
  format = c("RData", "tsv")
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació dels informes serà la subcarpeta \code{informes}.}

\item{etiquetes}{filtre d'etiquetes d'objectes d'OSM per la consulta d'Overpass seguint el format del paràmetre `features` d'[osmdata::add_osm_features()]. Combinació `O`.}

\item{claus}{filtre de claus d'objectes d'OSM per la consulta d'Overpass seguint el format del paràmetre `key` d'[osmdata::add_osm_feature()]. Combinació `I`.}

\item{actualitzaInformes}{si és `TRUE` i ja existeix el fitxer d'informe, el mou a la carpeta «ANTIC».}

\item{sufixFitxers}{text afegir com a sufix al nom dels fitxers dels informes (`arrelProjecte/informe-Regio-comarca$sufixFitxer$.tsv`).}

\item{divisions}{`data.frame` amb informació dels [PPCC], \code{\link{territoris}}, \code{\link{comarques}} o \code{\link{municipis}}. Per defecte, comarques.}

\item{consulta_wikidata}{Si és `TRUE`, afegeix les columnes `noms_wd`, `ca.viquipedia`, `wikidata_tipus` amb informació extreta de Wikidata pels objectes que tenen etiqueta wikidata.}

\item{coordenades}{Si és `TRUE`, els informes inclouen les columnes `latitude` i `longitude`. Per defecte, `FALSE`.}

\item{data}{Punt en el temps en format [ISO_8601](https://ca.wikipedia.org/wiki/ISO_8601). Per exemple, «2014-09-11T00:00:00Z» per l'11 de setembre de 2014. Si no s'especifica, consulta el moment actual.}

\item{data2}{Data en el mateix format que `data`. Si s'especifica, retorna la diferència entre `data` i `data2` (veure [wiki](https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#Difference_between_two_dates_(diff))).}

\item{adiff}{Si és `TRUE` i hi ha paràmetre `data(2)*`, afegeix columnes per les dades de [diferències augmentades](https://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL#Augmented-difference_between_two_dates_(adiff)).}

\item{format}{Com es desaran els resultats. Per defecte, "RData".}
}
\value{
Retorna una llista amb els informes.
}
\description{
Generar informes per territoris, comarques o municipis dels Països Catalans amb tots els objectes d'OSM segons els filtres especificats. Només inclou els municipis i zones catalanoparlants de les comarques bilingües.
}
\examples{
ordres <- obte_informesPPCC(
  arrelProjecte = "PPCC/osm_cat",
  sufixFitxers = "_name-osm_cat",
  divisions = comarques[!comarques$regio \%in\% c("CatNord", "Sardenya"), ]
)
}
