% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edicio.R
\name{prepara_edicions}
\alias{prepara_edicions}
\title{Genera fitxers amb edicions combinant els informes i revisions}
\usage{
prepara_edicions(
  arrelProjecte,
  informes,
  revisions,
  format = c("RData", "tsv", "osc", "xml")
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta `edicions` i
buscarà els informes i revisions a partir d'aquest camí si no hi ha paràmetres `informes` o `revisions`.}

\item{informes}{Camins dels informes o llista de les taules d'informes.}

\item{revisions}{Un `data.frame` amb revisions validades o un vector de camins de fitxers de revisions fetes.}

\item{format}{"RData", "tsv", "osc" o "xml". Serà l'extensió i format del fitxer a escriure.}
}
\value{
Una llista amb taules `osmapi_OsmChange` o `xml_document` per format `osc` o `xml`. Si `informes` conté els
  camins dels fitxers d'informe, s'escriuran els fitxers d'edicions a la carpeta `edicions/*.format` del projecte.
}
\description{
Fusiona les revisions fetes amb tots els informes que contenen objectes amb etiquetes iguals.
Reaprofita les revisions per casos que només es diferencien per si els noms comencen amb majúscules o minúscules,
mantenint el format original.
}
