% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processaReports.R
\name{preparaEdicions}
\alias{preparaEdicions}
\title{Genera fitxers amb edicions combinant els informes i revisions}
\usage{
preparaEdicions(arrelProjecte, usuari)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta \code{edicions} i buscarà els informes i revisions a partir d'aquest camí.}
}
\value{
Vector d'ordres per carregar els fitxers generats amb LangToolsOSM.
}
\description{
Reaprofita les revisions per casos que només es diferencien per si els noms comencen amb majúscules o minúscules, mantenint el format original.
}
