% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte.R
\name{recompteCasos}
\alias{recompteCasos}
\alias{recompteCasosInformes}
\alias{recompteCasosEdicions}
\title{Recompte de casos dels informes, edicions o ambdós}
\usage{
recompteCasosInformes(
  dades,
  divisions,
  arrelProjecte,
  campsUnics = c("name", "name:ca", "alt_name:ca", "alt_name", "noms_wd",
    "ca.viquipedia", "wikidata_tipus", "wikidata")
)

recompteCasosEdicions(arrelProjecte, edicions, divisions)

recompteCasos(arrelProjecte, informes, dades)
}
\arguments{
\item{dades}{un \code{data.frame} amb una columna anomenada \code{informe} (o \code{edicio} per la funció \code{recompteCasosEdicions}) amb els camins als fitxers «.tsv». Si s'especifica, s'ignoren la resta de paràmetres.}

\item{arrelProjecte}{camí a l'arrel del projecte. Es llegiran tots els fitxers *.tsv de la subcarpeta \code{informes} i/o \code{edicions/FET} si no s'especifiquen altres paràmetres.}

\item{edicions}{vector de caràcters amb els camins a fitxers d'edicions.}

\item{informes}{vector de caràcters amb els camins a fitxers d'informes.}
}
\value{
Una taula amb files per cada informe i amb les columnes següents:
  \describe{
    \item{informe}{camí del fitxer d'informe.}
    \item{nObjectes}{nombre d'objectes d'OSM.}
    \item{nCasos}{nombre de casos únics de les etiquetes \code{«name»}, \code{«name:ca»}, \code{«alt_name:ca»},
 \code{«alt_name»}, \code{«noms_wd»}, \code{«ca.viquipedia»} i \code{«wikidata»}.}
    \item{nObjectesNomWikidata}{nombre d'objectes d'OSM amb nom en català a wikidata (columna noms_wd dels informes).}
    \item{nCasosNomWikidata}{nombre de casos únics amb nom en català a wikidata (columna noms_wd dels fitxers de revisions).}
    \item{revisat}{si existeix o no unfitxer de revisió a la subcarpeta del projecte \code{revisions/FET}).}

  \strong{Només per la funció \code{recompteCasos}:}
    \item{nObjectesEditat}{nombre d'objectes editats d'OSM.}
    \item{nCasosEditat}{nombre de casos únics editats de les etiquetes \code{«name»}, \code{«name:ca»}, \code{«alt_name:ca»},
 \code{«alt_name»}, \code{«noms_wd»}, \code{«ca.viquipedia»} i \code{«wikidata»}.}
    \item{nObjectesNomWikidataEditat}{nombre d'objectes editats d'OSM amb nom en català a wikidata (columna noms_wd dels informes).}
    \item{nCasosNomWikidataEditat}{nombre de casos únics editats amb nom en català a wikidata (columna noms_wd dels fitxers de revisions).}
}
}
\description{
Compta el nombre de casos i casos únics, si tenen informació de wikidata o no
i altres detalls. Per les edicions, la informació s'agrega a nivell d'informe,
és a dir, si hi ha diversos fitxers d'edicions per un mateix informe, els suma.
}
\details{
S'assumeix que l'ordre dels infomres de dades `dades` coincideix amb les files de `divisions`.
}
