% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte_LangToolsOSM.R
\name{subdivisionsConsultaOverpass}
\alias{subdivisionsConsultaOverpass}
\title{Subdivisions d'àrees amb Overpass}
\usage{
subdivisionsConsultaOverpass(
  area,
  filtreSubdivisions,
  etiquetes = c("name", "'name:ca'", "wikidata"),
  format = c("csv", "json")
)
}
\arguments{
\item{area}{filtre de l'àrea per la que es vol generar les subdivisions.}

\item{filtreSubdivisions}{filtre per identificar les relacions que defineixen les subdivisions.}

\item{etiquetes}{etiquetes a afegir com a columnes a la sortida d'Overpass en el format \code{csv}.}

\item{format}{si és \code{csv}, la consulta retorna dades en aquest format. Si és \code{json}, mostra les divisions al mapa.}
}
\value{
Consulta per executar a \href{https://overpass-turbo.eu}{Overpass}.
}
\description{
Genera consultes per preparar taules amb les subdivisions d'àrees per \href{https://overpass-turbo.eu}{Overpass}.
Utilitzat per dividir països en àrees més manejables i per evitar conjunts de canvis massa grans a l'hora d'afegir exotopònims.
}
\examples{
consulta <- subdivisionsConsultaOverpass(
  area = "['name:ca'='Rússia'][admin_level=2]",
  filtreSubdivisions = "[admin_level=4]",
  etiquetes = c("name", "'name:ca'", "'wikidata'", "admin_level")
)
cat(consulta)
consulta <- subdivisionsConsultaOverpass(
  area = "['name:ca'='Rússia'][admin_level=2]",
  filtreSubdivisions = "[admin_level=4]",
  format = "json"
)
cat(consulta)
}
