% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informesPPCC.R
\name{generaInformesPPCC}
\alias{generaInformesPPCC}
\title{Generar informes per les zones catalanoparlants de les comarques dels PPCC}
\usage{
generaInformesPPCC(
  arrelProjecte,
  filtre,
  actualitzaInformes = FALSE,
  sufixFitxers = "",
  comarques = toponimsCat::comarques
)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació dels informes serà la subcarpeta \code{informes}.}

\item{filtre}{filtre d'etiquetes d'objectes d'OSM per la consulta d'Overpass.}

\item{actualitzaInformes}{si és \code{TRUE} i ja existeix el fitxer d'informe, el mou a la carpeta «ANTIC».}

\item{sufixFitxers}{text afegir com a sufix al nom dels fitxers dels informes («arrelProjecte/informe-Regio-comarca$sufixFitxer$.tsv»)}

\item{comarques}{\code{data.frame} amb informació de les comarques. Per defecte, \code{\link{comarques}}}
}
\value{
Retorna la taula de les comarques amb els nous camps \code{cmd}, que conté l'ordre per generar els informes amb LangTools,
i el camp \code{informe}, que conté el camí del fitxer de l'informe que es crearà.
}
\description{
Generar informes per les zones catalanoparlants de les comarques dels PPCC
}
\examples{
ordres<- generaInformesPPCC(arrelProjecte="ppcc/calle-carrer",
                            filtre="nwr[name~'^[Cc]alle'][!'name:ca']",
                            sufixFitxers="_name-calle")
\dontrun{
# Crida les ordres de LantToolsOSM (cal que estigui instal·lat a l'entorn Python configurat a R)
  for (i in 1:length(cmd)){
    message(i, " / ", length(cmd), "\t", cmd[i])
    system(ordres)
  }
}
}
