% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processaReports.R
\name{generaRevisions}
\alias{generaRevisions}
\alias{generaRevisions_regexName}
\alias{generaRevisions_regexTranslations}
\title{Genera fitxers de revisió d'informes}
\usage{
generaRevisions_regexName(
  informes,
  arrelProjecte = "ppcc/calle-carrer/",
  cerca = "([Cc])alle ",
  substitueix = "\\\\1arrer ",
  revisioUnificada = FALSE
)

generaRevisions_regexTranslations(
  informes,
  arrelProjecte = "exotopònims/Grècia/",
  cerca = " \\\\(.+\\\\)",
  substitueix = "",
  revisioUnificada = FALSE
)
}
\arguments{
\item{informes}{camí dels informes pels que es vol generar fitxers de revisió.}

\item{arrelProjecte}{camí a l'arrel del projecte. La carpeta de destinació serà la subcarpeta \code{revisions}.}

\item{cerca}{expressió regular del patró de cerca (\link{regex}).}

\item{substitueix}{text a substituir.}

\item{revisioUnificada}{si és \code{TRUE}, genera un unic fitxer de revisió. Sinó, genera un fixer de revisió per cada informe.}
}
\value{
Retorna els camíns dels fitxers de revisió generats.
}
\description{
Els fitxers de revisió d'informes contenen els casos a revisar amb combinacions úniques de
 les etiquetes \code{«name»}, \code{«name:ca»}, \code{«alt_name:ca»}, \code{«alt_name»},
 \code{«translations»}, \code{«ca.wikipedia_page»} i \code{«wikidata_id»}.
 Cal corregir els casos de \code{«name:ca»} i \code{«alt_name:ca»} incorrectes i esborrar-los o deixar-los en blanc si no és clar. Un cop revisat,
 moveu els fitxers de revisions a «arrelProjecte/revisio/FET/» i prepareu les edicions a OSM amb la funció \code{\link{preparaEdicions}}.
}
\seealso{
\link{gsub}
}
