% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projecte_LangToolsOSM.R
\name{recompteCasosInformes}
\alias{recompteCasosInformes}
\title{Recompte de casos dels informes}
\usage{
recompteCasosInformes(arrelProjecte, informes, dades)
}
\arguments{
\item{arrelProjecte}{camí a l'arrel del projecte. Es llegiran tots els fitxers *.tsv de la subcarpeta \code{informes} si no s'especifiquen altres paràmetres.}

\item{informes}{vector de caràcters amb els camins a fitxers d'informes.}

\item{dades}{un \code{data.frame} amb una columna anomenada \code{informe} amb els camins a fitxers d'informes. Si s'especifica, s'ignoren la resta de paràmetres.}
}
\value{
Una taula amb files per cada informe i amb les columnes següents:
  \describe{
    \item{nObjectes}{nombre d'objectes d'OSM.}
    \item{nCasos}{nombre de casos únics de les etiquetes \code{«name»}, \code{«name:ca»}, \code{«alt_name:ca»},
 \code{«alt_name»}, \code{«translations»}, \code{«ca.wikipedia_page»} i \code{«wikidata_id»}.}
    \item{revisat}{si existeix o no un fitxer de revisió a la subcarpeta del projecte \code{revisions/FET}).}
 }
}
\description{
Recompte de casos dels informes
}
