% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrector_ortografic.R
\name{revisionsSenseErrors}
\alias{revisionsSenseErrors}
\title{Revisions sense errors}
\usage{
revisionsSenseErrors(
  fitxersRevisions,
  sufix_nou = "_correcte.tsv",
  LanguageTool = FALSE,
  descartaErrorsLT = TRUE,
  disabled_rules = c("UPPERCASE_SENTENCE_START", "ESPAI_DARRERE_PUNTICOMA"),
  cl = 1
)
}
\arguments{
\item{fitxersRevisions}{camins de fitxers de revisions a processar.}

\item{sufix_nou}{sufix pels fitxers de revisió nous sense els casos amb errors.}

\item{LanguageTool}{si és \code{TRUE}, cerca errors també amb \code{\link[LanguageToolR]{languagetool}}. Pot ser força lent amb moltes dades.}

\item{descartaErrorsLT}{si és \code{TRUE}, per defecte, descarta els casos amb errors segons LanguageTool. Si és \code{FALSE}, afegeix informació de l'error a la taula (pot duplicar casos si contenen més d'un error).}

\item{disabled_rules}{identificadors de regles de LanguageTool a deshabilitar.}
}
\value{
Camins dels fitxers nous sense els casos en què \code{name:ca} o \code{alt_name:ca} tenen errors ortogràfics segons el diccionari hunspell (i LanguageTool).
}
\description{
Crea fitxers de revisions descartant els casos amb errors ortogràfics detectats per \code{\link[hunspell]{hunspell}}.
}
\details{
Quan el paràmetre \code{descartaErrorsLT} és \code{FALSE}, combina la taula d'informes amb la taula d'errors de LanguageTool.
Això fa que els casos que tenen més d'un error a l'etiqueta `name:ca` es repliquin.
La infromació dels errors del camp \code{alt_name:ca} no s'afegeix a la taula, només la columna \code{errorLT_alt_name:ca} amb \code{TRUE} o \code{FALSE}.
}
