\docType{package}
\name{toponimsCat-package}
\alias{toponimsCat}
\alias{toponimsCat-package}
\title{toponimsCat: Eines per ajudar a afegir l'etiqueta "name:ca" a OpenStreetMap}
\description{
Genera ordres per crear informes d'objectes d'OpenStreetMap amb LangToolsOSM a partir de filtres i àrees determinades, crea fitxers per revisar les modificacions i pujar els canvis.
}
\seealso{
Enllaços d'interès:
\itemize{
  \item \url{https://github.com/OSM-Catalan/toponimsCat}
  \item Report bugs at \url{https://github.com/OSM-Catalan/toponimsCat/issues}
  \item \href{https://github.com/OSM-Catalan/LangToolsOSM}{LangToolsOSM}
  \item \href{https://wiki.openstreetmap.org/wiki/WikiProject_Catalan/Topon\%C3\%ADmia_i_exotop\%C3\%B2nims}{Espai de coordinació per la toponímia de la comunitat d'OpenStreetMap Països Catalans}
  \item \href{https://www.openstreetmap.org}{OpenStreetMap}
}

}
\author{
\strong{Maintainer}: Joan Maspons \email{joanmaspons@gmail.com}

}
\keyword{internal}
